Put getJobDetailsFromUserInput and validateInput back
