+++ walk-in-interview/src/main/webapp/new-job/script.js
@@ -13,8 +13,7 @@ const CurrentLocale = 'en';
  */
 import {AppStrings} from '../strings.en.js';
 
-import {getRequirementsList, getJobDetailsFromUserInput, 
-  validateRequiredUserInput, setErrorMessage} from "../common-functions.js";
+import {getRequirementsList, setErrorMessage} from "../common-functions.js";
 
 const STRINGS = AppStrings['new-job'];
 const HOMEPAGE_PATH = '../index.html';
@@ -164,6 +163,115 @@ function renderJobExpiryLimits() {
   const datePicker = document.getElementById('expiry');
   datePicker.setAttribute('min', min);
   datePicker.setAttribute('max', max);
+}
+
+/**
+ * Gets job detail from user input.
+ * @return {Object} containing the user inputs.
+ */
+function getJobDetailsFromUserInput() {
+  const name = document.getElementById('title').value;
+  const description = document.getElementById('description').value;
+  const address = document.getElementById('address').value;
+  const postalCode = document.getElementById('postal-code').value;
+  const payFrequency = document.getElementById('pay-frequency').value;
+  const payMin = document.getElementById('pay-min').valueAsNumber;
+  const payMax = document.getElementById('pay-max').valueAsNumber;
+
+  const requirementsCheckboxes =
+      document.getElementsByName('requirements-list');
+  const requirementsList = [];
+  requirementsCheckboxes.forEach(({checked, id}) => {
+      if (checked) {
+      requirementsList.push(id);
+      }
+  });
+
+  const expiry = document.getElementById('expiry').valueAsNumber;
+  const duration = document.getElementById('duration').value;
+
+  const jobDetails = {
+      jobTitle: name,
+      jobLocation: {
+      address: address,
+      postalCode: postalCode,
+      lat: 1.3039, // TODO(issue/13): get these from places api
+      lon: 103.8358,
+      },
+      jobDescription: description,
+      jobPay: {
+      paymentFrequency: payFrequency,
+      min: payMin,
+      max: payMax,
+      },
+      requirements: requirementsList,
+      postExpiryTimestamp: expiry,
+      jobDuration: duration,
+  };
+
+  return jobDetails;
+}
+
+/**
+* Validates the user input
+* @return {boolean} depending on whether the input is valid or not.
+*/
+function validateRequiredUserInput() {
+  // TODO(issue/19): add more validation checks
+  const name = document.getElementById('title');
+  const description = document.getElementById('description');
+  const address = document.getElementById('address');
+  const postalCode = document.getElementById('postal-code');
+  const payFrequency = document.getElementById('pay-frequency').value;
+  const payMin = document.getElementById('pay-min').valueAsNumber;
+  const payMax = document.getElementById('pay-max').valueAsNumber;
+  const duration = document.getElementById('duration').value;
+  const expiry = document.getElementById('expiry').valueAsNumber;
+
+  if (name.value === '') {
+      setErrorMessage(/* msg */ name.placeholder,
+          /** includes default msg */ true);
+      return false;
+  }
+
+  if (description.value === '') {
+      setErrorMessage(/* msg */ description.placeholder,
+          /** includes default msg */ true);
+      return false;
+  }
+
+  if (address.value === '') {
+      setErrorMessage(/* msg */ address.placeholder,
+          /** includes default msg */ true);
+      return false;
+  }
+
+  if (postalCode.value === '') {
+      setErrorMessage(/* msg */ postalCode.placeholder,
+          /** includes default msg */ true);
+      return false;
+  }
+
+  if (payFrequency === '' || Number.isNaN(payMin) || Number.isNaN(payMax) ||
+      payMin > payMax || payMin < 0 || payMax < 0) {
+      setErrorMessage(/* msg */ document.getElementById('pay-title')
+          .textContent, /** includes default msg */ true);
+      return false;
+  }
+
+  if (duration === '') {
+      setErrorMessage(/* msg */ document.getElementById('duration-title')
+          .textContent, /** includes default msg */ true);
+      return false;
+  }
+
+  if (Number.isNaN(expiry)) {
+      setErrorMessage(/* msg */ document.getElementById('expiry-title')
+          .textContent, /** includes default msg */ true);
+      return false;
+  }
+
+  return true;
 }
 
 const submitButton = document.getElementById('submit');
